requirements.txt
# === Core Framework ===
Flask

# === Vector Store & Embeddings ===
faiss-cpu
sentence-transformers
scikit-learn
numpy

# === Document Ingestion ===
PyMuPDF
python-docx

# === Optional LLM API Clients ===
openai==0.28.1
anthropic==0.5.0
requests==2.31.0

# === Environment & System ===
python-dotenv

# === Dev & Testing ===
pytest
coverage

# === Frontend Utilities ===
Jinja2

# === Optional (safe build tools for native packages) ===
setuptools
wheel


# # === Core Framework ===
# Flask==2.3.2

# # === Vector Store & Embeddings ===
# faiss-cpu==1.7.4
# sentence-transformers==2.2.2
# scikit-learn==1.3.2
# numpy==1.24.4

# # === Document Ingestion ===
# PyMuPDF==1.23.3          # PDF parsing
# python-docx==1.0.1        # .docx parser

# # === Optional LLM API Clients ===
# openai==1.14.2            # OpenAI GPT integration
# anthropic==0.5.0          # Claude API integration
# requests==2.31.0          # Used for Claude and Ollama

# # === Environment & System ===
# python-dotenv==1.0.1      # Load from .env
# uuid==1.30                # Unique session IDs (used implicitly)

# # === Dev & Testing ===
# pytest==8.1.1
# coverage==7.4.4

# # === Frontend Utilities ===
# Jinja2==3.1.3             # Template engine (Flask default)


# fastapi
# uvicorn[standard]
# flask
# openai==0.28.1
# pydantic
# faiss-cpu
# sentence-transformers
# pytest
# flake8
# python-docx
# pandas
# pdfplumber
# httpx
# python-multipart
# python-dotenv